Crate bevy_winit

bevy_winit provides utilities to handle window creation and the eventloop through winit

Most commonly, the WinitPlugin is used as part of DefaultPlugins. The app’s runner is set by WinitPlugin and handles the winit EventLoop. See winit_runner for details.

Re-exports

pub use winit_event::*;

Modules

accessibility

Helpers for mapping window entities to accessibility types

winit_event

Structs

WakeUp

The default event that can be used to wake the window loop Wakes up the loop if in wait state

WinitPlugin

A Plugin that uses winit to create and manage windows, and receive window and input events.

WinitSettings

Settings for the WinitPlugin.

WinitWindows

A resource mapping window entities to their winit-backend Window states.

Enums

UpdateMode

Determines how frequently an App should update.

Functions

create_windows

Creates new windows on the winit backend for each entity with a newly-added Window component.

get_best_videomode

Gets the “best” video-mode handle from a monitor.

get_fitting_videomode

Gets the “best” video mode which fits the given dimensions.

winit_window_position

Compute the physical window position for a given WindowPosition.

Type Aliases

CreateWindowParams

The parameters of the create_windows system.

EventLoopProxy

The winit::event_loop::EventLoopProxy with the specific winit::event::Event::UserEvent used in the [winit_runner].