Struct bevy_winit::WinitSettings

pub struct WinitSettings {
    pub return_from_run: bool,
    pub focused_mode: UpdateMode,
    pub unfocused_mode: UpdateMode,
}

A resource for configuring usage of the rust_winit library.

Fields
return_from_run: bool

Configures winit to return control to the caller after exiting the event loop, enabling App::run() to return.

By default, return_from_run is false and Bevy will use winit’s EventLoop::run() to initiate the event loop. EventLoop::run() will never return but will terminate the process after the event loop exits.

Setting return_from_run to true will cause Bevy to use winit’s EventLoopExtRunReturn::run_return() instead which is strongly discouraged by the winit authors.

Supported platforms

This feature is only available on the following desktop target_os configurations: windows, macos, linux, dragonfly, freebsd, netbsd, and openbsd.

Setting return_from_run to true on unsupported platforms will cause App::run() to panic!

focused_mode: UpdateMode

Configures how the winit event loop updates while the window is focused.

unfocused_mode: UpdateMode

Configures how the winit event loop updates while the window is not focused.

Implementations

impl WinitSettings

pub fn game() -> Self

Configure winit with common settings for a game.

pub fn desktop_app() -> Self

Configure winit with common settings for a desktop application.

pub fn update_mode(&self, focused: bool) -> &UpdateMode

Gets the configured UpdateMode depending on whether the window is focused or not