Crate bevy_winit

bevy_winit provides utilities to handle window creation and the eventloop through winit

Most commonly, the WinitPlugin is used as part of DefaultPlugins. The app’s runner is set by WinitPlugin and handles the winit EventLoop. See winit_runner for details.

Modules

accessibility

Helpers for mapping window entities to accessibility types

Structs

WinitPlugin

A Plugin that utilizes winit for window creation and event loop management.

WinitSettings

A resource for configuring usage of the winit library.

WinitWindows

A resource which maps window entities to winit library windows.

Enums

UpdateMode

Configure how the winit event loop should update.

Functions

get_best_videomode

Gets the “best” videomode from a monitor.

get_fitting_videomode

Gets the “best” video mode which fits the given dimensions.

winit_runner

The default App::runner for the WinitPlugin plugin.

winit_window_position

Compute the physical window position for a given WindowPosition.