bevy_winit
Function create_windows

pub fn create_windows<F: QueryFilter + 'static>(
    event_loop: &ActiveEventLoop,
    (commands, created_windows, window_created_events, winit_windows, adapters, handlers, accessibility_requested, monitors): SystemParamItem<'_, '_, CreateWindowParams<'_, '_, F>>,
)

Creates new windows on the winit backend for each entity with a newly-added Window component.

If any of these entities are missing required components, those will be added with their default values.